Update pyfpgrowth.py
File has some missing lines of code. It does not give exact output before, after I changed few lines it is working fine. When I tried your code with the below example ,it is supposed to give (4,):3 ie. count of 4 is 3, but it does not.
transactions = [[1, 2, 5],
                [2, 4],
                [2, 3],
                [1, 2, 4],
                [1, 4],
                [2, 3],
                [1, 3],
                [1, 2, 3, 5],
                [1, 2, 3]]
patterns = pyfpgrowth.find_frequent_patterns(transactions, 2)